+++ packages/react/auto/src/index.ts
@@ -1,2 +1,0 @@
-import { installAutoSignalTracking } from "@preact/signals-react/runtime";
-installAutoSignalTracking();