+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @preact/signals-react
 
+## 3.0.1
+
+### Patch Changes
+
+- [#640](https://github.com/preactjs/signals/pull/640) [`503128f`](https://github.com/preactjs/signals/commit/503128f55af128ddc3b1f2133bc90f576d699d6c) Thanks [@andrewiggins](https://github.com/andrewiggins)! - Don't track signals read during an SSR render (e.g. renderToString, renderToStaticMarkup)
+
+## 3.0.0
+
+### Major Changes
+
+- [#638](https://github.com/preactjs/signals/pull/638) [`8c72dff`](https://github.com/preactjs/signals/commit/8c72dff85b1a6c8a2e724aeeab27daf95ad8c75f) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Remove `/auto` and automatic tracking as it is incompatible with React 19
+
 ## 2.3.0
 
 ### Minor Changes